Init rmsprop mean square state 'm' with 1 instead 0
With alpha near 1 (e.g. the default value 0.99) the gradient was
likely scaled up by a division by a number <1 during the first
few iterations.
With the original impl the learning rate had to be set to a much
smaller value when using rmsprop compared to plain-vanilla sgd in
order not to diverge.