Fixed the link to the Adam research paper
Fixed the link to the, "Adam: A Method for Stochastic Optimization" research paper. 

This link no longer works: http://arxiv.org/pdf/1412.6980.pdf

I and many others involved with machine learning, find it's better to link to the research paper's arXiv page itself, and not directly to the PDF file. This is because it's not easy to get to the research paper's arXiv page, directly from the PDF, but it is easy to get to the PDF from the arXiv page.